requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-toolbelt==0.0.1
gunicorn==20.0.4
image==1.5.33
numpy==1.20.1
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
